+++ iti-react-core/src/Hooks/index.ts
@@ -1,8 +1,9 @@
-﻿export * from './UseCancellablePromiseCleanup'
+﻿import { FieldValidity as _FieldValidity } from './UseFieldValidity'
+
+export * from './UseCancellablePromiseCleanup'
 export * from './UseValidationInProgressMonitor'
 export * from './UsePrevious'
 export * from './Query'
 
 export { fieldValidityIsValid, useFieldValidity } from './UseFieldValidity'
-import { FieldValidity as _FieldValidity } from './UseFieldValidity'
 export type FieldValidity = _FieldValidity