+++ iti-react-core/src/Util/GetGuid.ts
@@ -1,4 +1,4 @@
-﻿/* eslint-disable @typescript-eslint/no-explicit-any */
+﻿/* eslint-disable */
 
 // Source: https://gist.github.com/jed/982883
 // Great for React keys when there is no database ID that can be used