+++ iti-react-core/src/Validation/Hooks/UseValidation.ts
@@ -1,5 +1,5 @@
-﻿import React from 'react'
-import { useEffect } from 'react'
+﻿import React, { useEffect, useMemo, useRef } from 'react'
+import { noop } from 'lodash'
 import {
     Validator,
     getCombinedValidatorOutput,
@@ -7,7 +7,6 @@ import {
     combineValidatorOutput,
     AsyncValidator
 } from '../Validator'
-import { defaults } from 'lodash'
 import { useAsyncValidator } from './UseAsyncValidator'
 
 interface UseValidationPropOptions<TValue> {
@@ -16,8 +15,15 @@ interface UseValidationPropOptions<TValue> {
 
     validators: Validator<TValue>[]
 
-    // If you change the validators or asyncValidator, you must also change the validationKey.
-    // Otherwise, WithValidation has no way to know the validators have changed.
+    // Internally, `useValidation` calls `useMemo` on `validators` and `asyncValidator`
+    // since those props won't have stable identities if defined during the render
+    // as is typical. If `useValidation` did not do this, every component that rendered
+    // a validated component would have to call `useMemo` on `validators` and
+    // `asyncValidator` (or move the definitions outside of the component) to prevent
+    // infinite `useEffect` loops.
+    //
+    // If and when you `validators` or `asyncValidator` do change, you MUST pass a different
+    // `validationKey` for `useValidation` to pick up the changes.
     validationKey?: string | number
 
     asyncValidator?: AsyncValidator<TValue>
@@ -48,55 +54,49 @@ export interface UseValidationOutput {
     asyncValidationInProgress: boolean
 }
 
-export function useValidation<TValue>(
-    options: UseValidationOptions<TValue>
-): UseValidationOutput {
-    const {
-        value,
-        name,
-        onValidChange,
-        validators,
-        validationKey,
-        asyncValidator,
-        onAsyncError,
-        onAsyncValidationInProgressChange,
-        formLevelValidatorOutput
-    } = defaults(
-        { ...options },
-        {
-            onValidChange: () => {
-                /* no-op */
-            },
-            onAsyncError: () => {
-                /* no-op */
-            },
-            onAsyncValidationInProgressChange: () => {
-                /* no-op */
-            }
-        }
-    )
+export function useValidation<TValue>({
+    value,
+    name,
+    validationKey,
+    onAsyncError = noop,
+    formLevelValidatorOutput,
+    ...otherOptions
+}: UseValidationOptions<TValue>): UseValidationOutput {
+    /* eslint-disable react-hooks/exhaustive-deps */
+    const validators = useMemo(() => otherOptions.validators, [validationKey])
+    const asyncValidator = useMemo(() => otherOptions.asyncValidator, [validationKey])
+    /* eslint-enable react-hooks/exhaustive-deps */
 
     const synchronousValidatorOutput = getCombinedValidatorOutput(value, validators)
 
     const { asyncValidatorOutput, asyncValidationInProgress } = useAsyncValidator({
         value,
-        validationKey,
         synchronousValidatorsValid: synchronousValidatorOutput.valid,
         asyncValidator,
         onError: onAsyncError
     })
 
-    // onValidChange is not included in the dependencies array to prevent an infinite
-    // setState loop when the onValidChange function is defined in the render method.
+    const onValidChangeRef = useRef(otherOptions.onValidChange ?? noop)
     useEffect(() => {
-        onValidChange(
-            name,
-            synchronousValidatorOutput.valid && asyncValidatorOutput.valid
-        )
-    }, [name, synchronousValidatorOutput.valid, asyncValidatorOutput.valid])
+        onValidChangeRef.current = otherOptions.onValidChange ?? noop
+    })
+
+    const overallValid = synchronousValidatorOutput.valid && asyncValidatorOutput.valid
 
     useEffect(() => {
-        onAsyncValidationInProgressChange(name, asyncValidationInProgress)
+        onValidChangeRef.current(name, overallValid)
+    }, [name, overallValid])
+
+    const onAsyncValidationInProgressChangeRef = useRef(
+        otherOptions.onAsyncValidationInProgressChange ?? noop
+    )
+    useEffect(() => {
+        onAsyncValidationInProgressChangeRef.current =
+            otherOptions.onAsyncValidationInProgressChange ?? noop
+    })
+
+    useEffect(() => {
+        onAsyncValidationInProgressChangeRef.current(name, asyncValidationInProgress)
     }, [name, asyncValidationInProgress])
 
     const validatorOutputs = [synchronousValidatorOutput, asyncValidatorOutput]