+++ iti-react-core/src/Validation/index.ts
@@ -1,6 +1,6 @@
-export * from './Validator'
-export * from './Hooks'
-
 // Export the Validators module as an object
 import * as _Validators from './Validators'
+
+export * from './Validator'
+export * from './Hooks'
 export const Validators = _Validators