+++ iti-react-core/src/__tests__/Hooks/Query/UseParameterizedQuery.test.ts
@@ -1,0 +1,1 @@
+it('stable', () => fail())